add lib/storage.ts: migrate & validate realm json
