+++ src/lib/realm.ts
@@ -1,22 +1,20 @@
 import * as THREE from 'three';
 import * as Comlink from 'comlink';
-import localForage from 'localforage';
 
 import { Game } from './game';
 import { GameECS } from './gameECS';
 import { RealmWorker, ChunkGenerationResult } from './worker/realm';
+import { ExportedRealmJson, loadExportedRealm } from './storage';
 
-import { PackedObjRealmComponent, createBaseRealm, pack as packObjRealm, unpack as unpackObjRealm } from './obj/realm';
+import { createBaseRealm } from './obj/realm';
 import {
-  Cell, PackedChunkComponent, getChunkEntityComponents, createChunk, mergeChunk,
-  pack as packChunk, unpack as unpackChunk, destroy as destroyChunk,
+  Cell, getChunkEntityComponents, createChunk, mergeChunk, destroy as destroyChunk,
 } from './chunk/chunk';
 import { addChunkMeshToScene, removeChunkMeshFromScene } from './chunk/render';
-import { PackedSubObjComponent, pack as packSubObj, unpack as unpackSubObj } from './subObj/subObj';
 import { addOrRefreshSubObjToScene, destroySubObj } from './subObj/subObj';
 
-import { EntityRef, UUID, entityEqual } from './utils/ecs';
-import { Vec2, warnIfNotPresent, downloadJson } from './utils/utils';
+import { EntityRef, entityEqual } from './utils/ecs';
+import { Vec2, warnIfNotPresent } from './utils/utils';
 import { spawnWorker, listenToWorkerNextValue } from './utils/worker';
 import Map2D from './utils/map2d';
 
@@ -71,19 +69,6 @@ export function resetRealm(game: Game) {
   })();
 }
 
-export async function fetchRealm(realmObjUUID: UUID): Promise<ExportedRealmJson> {
-  const response = await fetch(`dev-objs/${realmObjUUID}-realm.json`);
-  if (warnIfNotPresent(response.ok)) return;
-  const json = await response.json() as ExportedRealmJson; // TODO: might need to be verified
-  if (realmObjUUID !== json.realmUUID) {
-    console.warn('UUID in json not equal');
-    return;
-  }
-
-  await localForage.setItem(`realm:${realmObjUUID}`, json);
-  return json;
-}
-
 export function switchRealm(json: ExportedRealmJson, game: Game) {
   const currentRealmObjComponents = game.ecs.getEntityComponents(game.realm.currentObj);
 
@@ -91,27 +76,6 @@ export function switchRealm(json: ExportedRealmJson, game: Game) {
   game.ecs.deallocate(game.realm.currentObj);
   game.realm.currentObj = loadExportedRealm(json, game.ecs);
   resetRealm(game);
-}
-
-export function loadExportedRealm(json: ExportedRealmJson, ecs: GameECS): EntityRef {
-  const newRealmEntity = ecs.fromUUID(json.realmUUID);
-  unpackObjRealm(newRealmEntity, json.packedObjRealm, ecs);
-  json.packedChunks.forEach(([UUID, packedChunk]) => {
-    unpackChunk(
-      ecs.fromUUID(UUID),
-      packedChunk,
-      ecs,
-    )
-  });
-  json.packedSubObjs.forEach(([UUID, packedSubObjs]) => {
-    unpackSubObj(
-      ecs.fromUUID(UUID),
-      packedSubObjs,
-      ecs,
-    );
-  });
-
-  return newRealmEntity;
 }
 
 function handleNextGeneratedChunk(result: ChunkGenerationResult, game: Game) {
@@ -170,51 +134,6 @@ function handleNextGeneratedChunk(result: ChunkGenerationResult, game: Game) {
 
 export function triggerRealmGeneration(centerChunkIJ: Vec2, game: Game) {
   game.realm.worker.triggerRealmGeneration(centerChunkIJ);
-}
-
-export interface ExportedRealmJson {
-  realmUUID: UUID;
-  packedObjRealm: PackedObjRealmComponent,
-  packedChunks: UUIDEntry<PackedChunkComponent>[],
-  packedSubObjs: UUIDEntry<PackedSubObjComponent>[],
-}
-type UUIDEntry<T> = [UUID, T];
-
-export function exportRealm(game: Game) {
-  const realmObjEntityComponents = game.ecs.getEntityComponents(game.realm.currentObj);
-
-  const realmUUID = game.ecs.getUUID(game.realm.currentObj);
-  const subObjUUIDs: UUID[] = [];
-
-  const packedObjRealm = packObjRealm(realmObjEntityComponents.get('obj/realm'), game.ecs);
-  const packedChunks = packedObjRealm.chunkEntries.map(([_chunkIJ, uuid]) => {
-    const chunk = game.ecs.getComponent(game.ecs.fromUUID(uuid), 'chunk');
-    const packedChunk = packChunk({
-      ...chunk,
-      subObjs: chunk.subObjs.filter(subObjEntity => !entityEqual(subObjEntity, game.player.subObjEntity))
-    }, game.ecs);
-    subObjUUIDs.push(...packedChunk.subObjs);
-
-    return [uuid, packedChunk] as UUIDEntry<PackedChunkComponent>;
-  });
-  const packedSubObjs = subObjUUIDs.map(uuid => ([
-    uuid,
-    packSubObj(
-      game.ecs.getComponent(
-        game.ecs.fromUUID(uuid),
-        'subObj',
-      ),
-      game.ecs,
-    )
-  ] as UUIDEntry<PackedSubObjComponent>));
-  
-  const objRealmJson: ExportedRealmJson = {
-    realmUUID,
-    packedObjRealm,
-    packedChunks,
-    packedSubObjs,
-  };
-  downloadJson(objRealmJson, `${realmUUID}-realm.json`);
 }
 
 function createBaseMaterial(): THREE.Material {
